Start point to work on remaining entites
